requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyarrow>=14.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
httpx[socks]>=0.27.0
socksio>=1.0.0
